pyproject.toml
[tool.poetry]
name = "memento-backend"
version = "0.1.0"
description = ""
authors = ["NoahCardoza <noahcardoza@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.22"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
